pyproject.toml
[project]
name = "sca-scripts"
version = "0.1.0"
description = "Useful scripts for SCA"
authors = [
    {name = "Vsevolod Rakita"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
SCA_scripts = "SCA_scripts.main:main"
